# [go] KeyCache Locking Improvements