--- a/go/appencryption/key_cache.go
+++ b/go/appencryption/key_cache.go
@@ -173,9 +173,19 @@ func isReloadRequired(entry cacheEntry, checkInterval time.Duration) bool {
 // is not present in the cache it will retrieve the key using the provided loader
 // and store the key if an error is not returned.
 func (c *keyCache) GetOrLoad(id KeyMeta, loader func(KeyMeta) (*internal.CryptoKey, error)) (*cachedCryptoKey, error) {
+	c.rw.RLock()
+	k, ok := c.getFresh(id)
+	c.rw.RUnlock()
+
+	if ok {
+		return tracked(k), nil
+	}
+
 	c.rw.Lock()
 	defer c.rw.Unlock()
 
+	// exit early if the key doesn't need to be reloaded just in case it has
+	// been loaded by rw lock in front of us
 	if k, ok := c.getFresh(id); ok {
 		return tracked(k), nil
 	}